refactor functions in pyfunc to models.utils as validate_schema depends on numpy and pands
